[com_publications] Squashed commit of the following...
commit d93fc5fdb3ad023bf3c1003528cc41e3eaa4e931
Author: M. Drew LaMar <mdlama@wm.edu>
Date:   Fri Aug 31 12:13:23 2018 -0400

    [PURR][#1862]Fix whitespaces

    (cherry picked from commit 54657a4232225360c00b1e64f60060c5fe617311)

commit a071650929b57066c4c3d42ffe598f80e29ae29a
Author: M. Drew LaMar <mdlama@wm.edu>
Date:   Fri Aug 31 12:12:39 2018 -0400

    [PURR][#1862]Create Entry functions for DOI register and update process

    (cherry picked from commit 56d9ffcd2c3ff9f174314fd9e74aeca56d4da6bb)

commit f2e3cf2641928c4730232379cf5d727ac2c12cb3
Author: root <root@example.com>
Date:   Wed Aug 22 10:17:17 2018 -0400

    [PURR][#1862]Fix the styling error

    (cherry picked from commit 6b91fc7d7ff83f05f9a34945a4bf3026419f581a)

commit ec0880965679934358b228dd8dfd9cc5ff537fcf
Author: M. Drew LaMar <mdlama@wm.edu>
Date:   Fri Aug 31 12:11:21 2018 -0400

    [PURR][#1862]Fix the styling errors

    (cherry picked from commit dc60301894c70aa29a4e5a4b043716ea83d77806)

commit a5813527988dc8b9864af3c11d21c3f433127b77
Author: M. Drew LaMar <mdlama@wm.edu>
Date:   Fri Aug 31 12:10:43 2018 -0400

    [PURR][#1862]DataCite MDS API application on PURR/Hubzero

    (cherry picked from commit 99ca9f86cb7be29d3250f9daf10c5645774ca623)

commit bc6adb91a8f28a7a62280ea4cf5fa344f9c58c1f
Author: M. Drew LaMar <mdlama@wm.edu>
Date:   Fri Aug 31 12:06:17 2018 -0400

    [PURR][#1862]Update to the configuration options

    (cherry picked from commit 953a38da4d485713eb18260da3127662989d7f38)

commit 192a52cba3fdee9549f679b6d14d4c393c802a77
Author: M. Drew LaMar <mdlama@wm.edu>
Date:   Fri Aug 31 12:05:12 2018 -0400

    [PURR][#1862]Use list options to switch from EZID to DataCite service

    (cherry picked from commit 14bdb9ef1ae4bd1206ab332d35fa91ed93a830e0)

commit 3463505963b087db6b611a4f4de59469ae356886
Author: M. Drew LaMar <mdlama@wm.edu>
Date:   Fri Aug 31 12:03:05 2018 -0400

    [PURR][#1862]Fix styling issue

    (cherry picked from commit 9857ff123fbdd890528989b2c6c9fccf9c35b410)

commit a2db4a02f26a5383ac0d2adf53a3772b084d697c
Author: root <root@example.com>
Date:   Mon Aug 13 13:10:53 2018 -0400

    [PURR][#1862]Fix the styling issue

    (cherry picked from commit 70aef3a9629091b066ab8264512e7beba00b81c6)

commit d6d5520a8c8ac672afb5f9e2eaa086caf8b6cf60
Author: M. Drew LaMar <mdlama@wm.edu>
Date:   Fri Aug 31 12:02:03 2018 -0400

    [PURR][#1862]Add switch control to EZID and DataCite DOI services

    (cherry picked from commit 399aec54143cf72568ffd808da0fcf060eed8967)

commit 9177f3841850ae3f748a81c0aa98181d258d7f18
Author: M. Drew LaMar <mdlama@wm.edu>
Date:   Fri Aug 31 12:01:26 2018 -0400

    [PURR][#1862]DataCite MDS API application on PURR/Hubzero

    (cherry picked from commit 5caf160ace68dd0d7659c532e591ab3f6829c825)

commit b24b943cf9a73547e0ec359b5cb0a4a0e702b640
Author: M. Drew LaMar <mdlama@wm.edu>
Date:   Fri Aug 31 12:00:45 2018 -0400

    [PURR][#1862]DataCite MDS API application on PURR/Hubzero

    (cherry picked from commit b7eb123801c3dac683935403de7586eeafb10d59)

commit 6b010275d253b5dcf5d1ac23dbec403767c7c012
Author: M. Drew LaMar <mdlama@wm.edu>
Date:   Fri Aug 31 11:59:22 2018 -0400

    [fix] Change publication ftp bundle symlink to reference relative paths.
    Absolute paths to the original publication file were not accessible via FTP.
    I changed the symlinks to reference the relative path instead.

    refs: https://purr.purdue.edu/support/ticket/1852
    (cherry picked from commit 19b920f2cdf4b1d46f93cb1c2356d4e39e701f19)

commit 50c14c03c61563e48d678037e9ccf1a493485b6d
Author: M. Drew LaMar <mdlama@wm.edu>
Date:   Fri Aug 31 11:57:35 2018 -0400

    [fix] Only attempt to remove a symlink file if it already exists

    New publication submissions were displaying an error since it was attempting to
    remove symlink that wasn't created yet.

    refs: https://purr.purdue.edu/support/ticket/1879
    (cherry picked from commit 3fe519396f39a367207a85c428e7b85e76071a75)

commit 9aac9ec3b323d807f32f501ad3f3225b7bdf693b
Author: M. Drew LaMar <mdlama@wm.edu>
Date:   Fri Aug 31 11:53:50 2018 -0400

    Add the layout class to all the aside-containing views to indicate such a layout to better target with CSS

    (cherry picked from commit 586a027aa8c6d9541417eceb36ba36e23c854ebb)

commit 7b5630e0f20b57068c761959d5c2b41a0d6e4c2e
Author: M. Drew LaMar <mdlama@wm.edu>
Date:   Fri Aug 31 11:51:24 2018 -0400

    [maint] Removing unneeded JS, standardizing calls

    (cherry-picked from e1e039c1a6)

commit 076d3f99b208157470825582d7a65e42c62b0ef5
Author: M. Drew LaMar <mdlama@wm.edu>
Date:   Fri Aug 31 11:47:49 2018 -0400

    [fix] Add check to publication request review

    Fixes: https://qubeshub.org/support/ticket/1205
    (cherry picked from commit f4594e59d875e941751658da842f623ee527d6e4)

...

+++ app/components/com_publications/models/curation.php
@@ -2158,8 +2158,13 @@ class Curation extends Obj
 	public function createSymLink()
 	{
 		$tarname = $this->getBundleName();
-		$tarpath = $this->_pub->path('base', true) . DS . $tarname;
+		$tarpath = $this->_pub->path('relative') . DS . $tarname;
 		$symLink = $this->_symLinkPath();
+		if ($symLink !== false)
+		{
+			chdir(dirname($symLink));
+		}
+
 		if (empty($this->_pub) || $symLink == false || !is_file($tarpath))
 		{
 			return false;
@@ -2186,7 +2191,11 @@ class Curation extends Obj
 		{
 			return false;
 		}
-		unlink($symLink);
+
+		if (is_file($symLink))
+		{
+			unlink($symLink);
+		}
 		return true;
 	}
